+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add function to get the name of a stem of a file
+
 ## [0.11.0] - 2020-03-26
 ### Added
 - Add function to merge lists together